Fixing fbprophet adapter
