+++ .binder/requirements.txt
@@ -1,15 +1,16 @@
 catch22
 cython>=0.29.0
-fbprophet
+fbprophet>=0.7.1
 hcrystalball>=0.1.9
 matplotlib
 notebook
 numba==0.50.*
-pandas>=1.0.0
+pandas>=1.1.0
 pmdarima>=1.8.0
 scikit-learn==0.23.*
 scikit-posthocs
 seaborn
 statsmodels>=0.12.1
-tbats
-tsfresh
+stumpy>=1.5.1
+tbats>=1.1.0
+tsfresh>=0.17.0
