+++ sktime/forecasting/base/adapters/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3 -u
+# -*- coding: utf-8 -*-
+
+__author__ = ["Markus Löning"]
+__all__ = ["_ProphetAdapter", "_StatsModelsAdapter", "_TbatsAdapter"]
+
+from sktime.forecasting.base.adapters._fbprophet import _ProphetAdapter
+from sktime.forecasting.base.adapters._statsmodels import _StatsModelsAdapter
+from sktime.forecasting.base.adapters._tbats import _TbatsAdapter
