+++ sktime/forecasting/bats.py
@@ -6,7 +6,7 @@ __author__ = ["Martin Walter"]
 __all__ = ["BATS"]
 
 from sktime.utils.validation._dependencies import _check_soft_dependencies
-from sktime.forecasting.base._adapters import _TbatsAdapter
+from sktime.forecasting.base.adapters import _TbatsAdapter
 
 _check_soft_dependencies("tbats")
