+++ sktime/forecasting/ets.py
@@ -2,7 +2,7 @@
 __all__ = ["AutoETS"]
 __author__ = ["Hongyi Yang"]
 
-from sktime.forecasting.base._adapters import _StatsModelsAdapter
+from sktime.forecasting.base.adapters import _StatsModelsAdapter
 from statsmodels.tsa.exponential_smoothing.ets import ETSModel as _ETSModel
 from itertools import product
 from joblib import delayed, Parallel
