+++ sktime/forecasting/tbats.py
@@ -5,7 +5,7 @@
 __author__ = ["Martin Walter"]
 __all__ = ["TBATS"]
 
-from sktime.forecasting.base._adapters import _TbatsAdapter
+from sktime.forecasting.base.adapters import _TbatsAdapter
 from sktime.utils.validation._dependencies import _check_soft_dependencies
 
 _check_soft_dependencies("tbats")
