+++ sktime/utils/_testing/estimator_checks.py
@@ -23,7 +23,7 @@ from sklearn.utils.estimator_checks import (
     check_get_params_invariance as _check_get_params_invariance,
 )
 from sklearn.utils.estimator_checks import check_set_params as _check_set_params
-from sklearn.utils.testing import set_random_state
+from sklearn.utils._testing import set_random_state
 from sklearn.utils.validation import check_random_state
 
 from sktime.base import BaseEstimator
@@ -511,29 +511,21 @@ def _get_err_msg(estimator):
 
 def _construct_instance(Estimator):
     """Construct Estimator instance if possible"""
+    # if non-default parameters are required, but none have been found,
+    # raise error
+    if hasattr(Estimator, "_required_parameters"):
+        required_parameters = getattr(Estimator, "required_parameters", [])
+        if len(required_parameters) > 0:
+            raise ValueError(
+                f"Estimator: {Estimator} requires "
+                f"non-default parameters for construction, "
+                f"but none have been found"
+            )
 
-    # construct with parameter configuration for testing
-    if Estimator in ESTIMATOR_TEST_PARAMS:
-        params = ESTIMATOR_TEST_PARAMS[Estimator]
-        estimator = Estimator(**params)
-
-    # otherwise construct with default parameters
-    else:
-        # if non-default parameters are required, but none have been found,
-        # raise error
-        if hasattr(Estimator, "_required_parameters"):
-            required_parameters = getattr(Estimator, "required_parameters", [])
-            if len(required_parameters) > 0:
-                raise ValueError(
-                    f"Estimator: {Estimator} requires "
-                    f"non-default parameters for construction, "
-                    f"but none have been found"
-                )
-
-        # construct with default parameters if none are required
-        estimator = Estimator()
-
-    return estimator
+    # construct with parameter configuration for testing, otherwise construct with
+    # default parameters (empty param dict)
+    params = ESTIMATOR_TEST_PARAMS.get(Estimator, {})
+    return Estimator(**params)
 
 
 def _make_args(estimator, method, **kwargs):
